--- scapture/src/main/java/com/server/scapture/user/JwtUtil.java
@@ -1,72 +1,0 @@
-package com.server.scapture.user;
-
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.util.Date;
-
-@Slf4j
-@Component
-public class JwtUtil {
-    @Value("${jwt.secret}")
-    private String SECRET_KEY;
-
-    private SecretKey getSigningKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(this.SECRET_KEY);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    // 액세스 토큰을 발급하는 메서드
-    public String generateAccessToken(Long userId, long expirationMillis) {
-        log.info("액세스 토큰이 발행되었습니다.");
-
-        return Jwts.builder()
-                .claim("userId", userId.toString()) // 클레임에 userId 추가
-                .setIssuedAt(new Date())
-                .setExpiration(new Date(System.currentTimeMillis() + expirationMillis))
-                .signWith(this.getSigningKey())
-                .compact();
-    }
-
-    // 응답 헤더에서 액세스 토큰을 반환하는 메서드
-    public String getTokenFromHeader(String authorizationHeader) {
-        return authorizationHeader.substring(7);
-    }
-
-    // 토큰에서 유저 id를 반환하는 메서드
-    public String getUserIdFromToken(String token) {
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(this.getSigningKey())
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody()
-                    .get("userId", String.class);
-        } catch (JwtException | IllegalArgumentException e) {
-            log.warn("유효하지 않은 토큰입니다.");
-            return null;
-        }
-    }
-
-    // Jwt 토큰의 유효기간을 확인하는 메서드
-    public Boolean isTokenExpired(String token) {
-        try {
-            Date expirationDate = Jwts.parserBuilder()
-                    .setSigningKey(this.getSigningKey())
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody()
-                    .getExpiration();
-            return expirationDate.before(new Date());
-        } catch (JwtException | IllegalArgumentException e) {
-            log.warn("유효하지 않은 토큰입니다.");
-            return null;
-        }
-    }
-}
