--- scapture/src/main/java/com/server/scapture/user/service/CustomOAuth2UserService.java
@@ -1,7 +1,0 @@
-package com.server.scapture.user.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CustomOAuth2UserService {
-}
