+++ scapture/src/main/java/com/server/scapture/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.server.scapture.user.service;
+
+public interface UserService {
+}
