+++ scapture/src/main/java/com/server/scapture/user/service/UserServiceImpl.java
@@ -1,0 +1,17 @@
+package com.server.scapture.user.service;
+
+import com.server.scapture.user.repository.UserRepository;
+
+import lombok.RequiredArgsConstructor;
+
+import org.springframework.stereotype.Service;
+
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService{
+
+    private final UserRepository userRepository;
+
+
+}
